refactor: adds and corrects types of ContentstackCollectin and marketplace functions
