+++ types/marketplace/app/oath.d.ts
@@ -1,0 +1,8 @@
+import { AnyProperty } from '../../utility/fields';
+import { AppOAuth } from '.';
+
+export interface Oauth {
+  fetch(param?: AnyProperty): Promise<AppOAuth>
+  update(data: { config: AppOAuth, param?: AnyProperty }): Promise<AppOAuth>
+  getScopes(): Promise<AnyProperty>
+}
