+++ types/marketplace/installation.d.ts
@@ -7,11 +7,16 @@ export interface Installation extends SystemFields {
     uninstall(param?: AnyProperty): Promise<AnyProperty>
     configuration(param?: AnyProperty): Promise<AnyProperty>
     setConfiguration(config: AnyProperty): Promise<AnyProperty>
+    getConfigLocation(): Promise<AnyProperty>
     serverConfig(param?: AnyProperty): Promise<AnyProperty>
     setServerConfig(config: AnyProperty): Promise<AnyProperty>
     installationData(): Promise<AnyProperty>
+    webhooks(webhookUid: string): WebHooks
 }
 
 export interface Installations {
-    findAll(param?: AnyProperty): Promise<ContentstackCollection<Installation>>
+    fetchAll(param?: AnyProperty): Promise<ContentstackCollection<Installation>>
+    getInstalledApps(): Promise<AnyProperty>
+    getInstalledUsers(): Promise<AnyProperty>
+    getInstalledStacks(): Promise<AnyProperty>
 }
