+++ types/marketplace/webhooks.d.ts
@@ -1,0 +1,7 @@
+import { AnyProperty } from "../utility/fields";
+
+export interface WebHooks {
+  listExecutionLogs(): Promise<AnyProperty>
+  getExecutionLog(executionUid: string): Promise<AnyProperty>
+  retryExecution(executionUid: string): Promise<AnyProperty>
+}
